fix: split test in multiple files and change last assertion
